requirements.txt
asgiref==3.2.7
Django==3.0.5
django-phone-field==1.8.0
djangorestframework==3.11.0
python-dotenv==0.12.0
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
